ccl: move pkg/benchccl to  pkg/bench
Note that I added a shard to account for the additional expectation
test that runs for the multiregion registry. I left the size as large
despite rttanalysisccl being marked as enormous based on a few local
runs.

Informs #164337
Release note: None
