+++ pkg/BUILD.bazel
@@ -23,7 +23,6 @@ ALL_TESTS = [
     "//pkg/build/starlarkutil:starlarkutil_test",
     "//pkg/build/util:util_test",
     "//pkg/build:build_test",
-    "//pkg/ccl/benchccl/rttanalysisccl:rttanalysisccl_test",
     "//pkg/ccl/changefeedccl/avro:avro_test",
     "//pkg/ccl/changefeedccl/cdceval:cdceval_test",
     "//pkg/ccl/changefeedccl/cdcevent:cdcevent_test",
@@ -975,7 +974,6 @@ GO_TARGETS = [
     "//pkg/build/util:util_test",
     "//pkg/build:build",
     "//pkg/build:build_test",
-    "//pkg/ccl/benchccl/rttanalysisccl:rttanalysisccl_test",
     "//pkg/ccl/changefeedccl/avro:avro",
     "//pkg/ccl/changefeedccl/avro:avro_test",
     "//pkg/ccl/changefeedccl/cdceval:cdceval",
