+++ pkg/bench/rttanalysis/BUILD.bazel
@@ -46,6 +46,7 @@ go_test(
         "grant_revoke_bench_test.go",
         "grant_revoke_role_bench_test.go",
         "jobs_test.go",
+        "multi_region_bench_test.go",
         "orm_queries_bench_test.go",
         "rtt_analysis_test.go",
         "system_bench_test.go",
@@ -58,9 +59,12 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":rttanalysis"],
     exec_properties = {"test.Pool": "large"},
-    shard_count = 4,
+    shard_count = 5,
     deps = [
         "//pkg/base",
+        "//pkg/ccl/multiregionccl",
+        "//pkg/ccl/multiregionccl/multiregionccltestutils",
+        "//pkg/ccl/partitionccl",
         "//pkg/jobs/jobspb",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
