+++ pkg/bench/rttanalysis/testdata/benchmark_expectations
@@ -1,7 +1,22 @@
 exp,benchmark
+14,"Discard/DISCARD_ALL,_1_tables_in_1_db"
+19,"Discard/DISCARD_ALL,_2_tables_in_2_dbs"
+0,"Discard/DISCARD_ALL,_no_tables"
+12-14,AlterPrimaryRegion/alter_empty_database_alter_primary_region
+16,AlterPrimaryRegion/alter_empty_database_set_initial_primary_region
+12-14,AlterPrimaryRegion/alter_populated_database_alter_primary_region
+18,AlterPrimaryRegion/alter_populated_database_set_initial_primary_region
+11,AlterRegions/alter_empty_database_add_region
+12-16,AlterRegions/alter_empty_database_drop_region
+13,AlterRegions/alter_populated_database_add_region
+16,AlterRegions/alter_populated_database_drop_region
 12,AlterRole/alter_role_with_1_option
 15,AlterRole/alter_role_with_2_options
 19,AlterRole/alter_role_with_3_options
+12,AlterSurvivalGoals/alter_empty_database_from_region_to_zone
+12,AlterSurvivalGoals/alter_empty_database_from_zone_to_region
+24,AlterSurvivalGoals/alter_populated_database_from_region_to_zone
+24,AlterSurvivalGoals/alter_populated_database_from_zone_to_region
 14,AlterTableAddCheckConstraint/alter_table_add_1_check_constraint
 14,AlterTableAddCheckConstraint/alter_table_add_2_check_constraints
 14,AlterTableAddCheckConstraint/alter_table_add_3_check_constraints
@@ -21,6 +36,12 @@ exp,benchmark
 13,AlterTableDropConstraint/alter_table_drop_1_check_constraint
 13,AlterTableDropConstraint/alter_table_drop_2_check_constraints
 13,AlterTableDropConstraint/alter_table_drop_3_check_constraints
+18,AlterTableLocality/alter_from_global_to_rbr
+10,AlterTableLocality/alter_from_global_to_regional_by_table
+10,AlterTableLocality/alter_from_rbr_to_global
+10,AlterTableLocality/alter_from_rbr_to_regional_by_table
+10,AlterTableLocality/alter_from_regional_by_table_to_global
+18,AlterTableLocality/alter_from_regional_by_table_to_rbr
 8,AlterTableSplit/alter_table_split_at_1_value
 11,AlterTableSplit/alter_table_split_at_2_values
 14,AlterTableSplit/alter_table_split_at_3_values
@@ -32,9 +53,6 @@ exp,benchmark
 16,CreateRole/create_role_with_2_options
 19,CreateRole/create_role_with_3_options
 14,CreateRole/create_role_with_no_options
-14,"Discard/DISCARD_ALL,_1_tables_in_1_db"
-19,"Discard/DISCARD_ALL,_2_tables_in_2_dbs"
-0,"Discard/DISCARD_ALL,_no_tables"
 15,DropDatabase/drop_database_0_tables
 16,DropDatabase/drop_database_1_table
 16,DropDatabase/drop_database_2_tables
@@ -51,9 +69,9 @@ exp,benchmark
 14,DropView/drop_1_view
 15,DropView/drop_2_views
 15,DropView/drop_3_views
-6,GenerateObjects/generate_1000_tables_-_this_test_should_use_the_same_number_of_RTTs_as_for_10_tables
-9,GenerateObjects/generate_100_tables_from_template
 6,GenerateObjects/generate_10_tables
+9,GenerateObjects/generate_100_tables_from_template
+6,GenerateObjects/generate_1000_tables_-_this_test_should_use_the_same_number_of_RTTs_as_for_10_tables
 17,GenerateObjects/generate_10x10_schemas_and_tables_in_existing_db
 6,GenerateObjects/generate_50000_tables
 8,Grant/grant_all_on_1_table
@@ -73,6 +91,9 @@ exp,benchmark
 9-12,Jobs/resume_job
 3,Jobs/show_job
 3-5,Jobs/show_jobs
+10,MultiRegionVirtualTableQueries/select_from_crdb_internal.zones_(10_tables)
+10,MultiRegionVirtualTableQueries/select_from_crdb_internal.zones_(100_tables)
+10,MultiRegionVirtualTableQueries/select_from_crdb_internal.zones_(50_tables)
 3,ORMQueries/activerecord_type_introspection_query
 3-4,ORMQueries/asyncpg_types
 5,ORMQueries/column_descriptions_json_agg
@@ -130,8 +151,8 @@ exp,benchmark
 17,Truncate/truncate_2_column_1_rows
 17,Truncate/truncate_2_column_2_rows
 0,UDFResolution/select_from_udf
-25,UseManyRoles/use_25_roles
 2,UseManyRoles/use_2_roles
+25,UseManyRoles/use_25_roles
 3,VirtualTableQueries/capture_index_usage_stats
 2,VirtualTableQueries/select_crdb_internal.invalid_objects_with_1_fk
 1,VirtualTableQueries/select_crdb_internal.tables_with_1_fk
