--- pkg/ccl/benchccl/rttanalysisccl/BUILD.bazel
@@ -1,26 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "rttanalysisccl_test",
-    size = "enormous",
-    srcs = [
-        "bench_test.go",
-        "multi_region_bench_test.go",
-    ],
-    data = glob(["testdata/**"]),
-    exec_properties = {"test.Pool": "large"},
-    shard_count = 16,
-    deps = [
-        "//pkg/base",
-        "//pkg/bench/rttanalysis",
-        "//pkg/ccl",
-        "//pkg/ccl/multiregionccl/multiregionccltestutils",
-        "//pkg/security/securityassets",
-        "//pkg/security/securitytest",
-        "//pkg/server",
-        "//pkg/testutils/pgurlutils",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/testcluster",
-        "//pkg/util/randutil",
-    ],
-)
