--- pkg/ccl/benchccl/rttanalysisccl/testdata/benchmark_expectations
@@ -1,22 +1,0 @@
-exp,benchmark
-12-14,AlterPrimaryRegion/alter_empty_database_alter_primary_region
-16,AlterPrimaryRegion/alter_empty_database_set_initial_primary_region
-12-14,AlterPrimaryRegion/alter_populated_database_alter_primary_region
-18,AlterPrimaryRegion/alter_populated_database_set_initial_primary_region
-11,AlterRegions/alter_empty_database_add_region
-12-16,AlterRegions/alter_empty_database_drop_region
-13,AlterRegions/alter_populated_database_add_region
-16,AlterRegions/alter_populated_database_drop_region
-12,AlterSurvivalGoals/alter_empty_database_from_region_to_zone
-12,AlterSurvivalGoals/alter_empty_database_from_zone_to_region
-24,AlterSurvivalGoals/alter_populated_database_from_region_to_zone
-24,AlterSurvivalGoals/alter_populated_database_from_zone_to_region
-18,AlterTableLocality/alter_from_global_to_rbr
-10,AlterTableLocality/alter_from_global_to_regional_by_table
-10,AlterTableLocality/alter_from_rbr_to_global
-10,AlterTableLocality/alter_from_rbr_to_regional_by_table
-10,AlterTableLocality/alter_from_regional_by_table_to_global
-18,AlterTableLocality/alter_from_regional_by_table_to_rbr
-10,VirtualTableQueries/select_from_crdb_internal.zones_(100_tables)
-10,VirtualTableQueries/select_from_crdb_internal.zones_(10_tables)
-10,VirtualTableQueries/select_from_crdb_internal.zones_(50_tables)
